requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.43.0
Django==5.1.7
django-cloudinary-storage==0.3.0
djangorestframework==3.15.2
idna==3.10
Pillow==8.2.0
pytz==2025.1
requests==2.32.3
setuptools==77.0.3
six==1.17.0
sqlparse==0.5.3
urllib3==2.3.0
wheel==0.45.1
